requirements.txt
paho.mqtt
coverage
isort
black
pylint
build
sphinx-rtd-theme
sphinx-pydantic
myst_parser
twine
